requirements.txt
PyQt5>=5.15.0
scapy>=2.4.5
requests>=2.25.1
cryptography>=3.4.7
psutil>=5.8.0
python-nmap>=0.7.1
dnspython>=2.1.0
beautifulsoup4>=4.9.3
pyopenssl>=20.0.1
pillow>=8.2.0
argparse>=1.4.0
whois>=0.9.15
tldextract>=3.3.0
validators>=0.20.0
urllib3>=1.26.7
colorama>=0.4.4
tabulate>=0.8.9
rich>=12.0.0
pyperclip>=1.8.2
zxcvbn-python>=4.4.24
python-whois>=0.7.3
lxml>=4.8.0
packaging>=21.3
